# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Farm Manager API
  version: 0.1.0
  description: API de gestion d'exploitation avicole (poulets de chair)
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /financement:
    get:
      operationId: listFinancements
      tags: [financement]
      summary: Liste des financements
      responses:
        "200":
          description: Liste des financements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Financement"
    post:
      operationId: createFinancement
      tags: [financement]
      summary: Créer un financement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFinancementBody"
      responses:
        "201":
          description: Financement créé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Financement"
  /financement/{id}:
    put:
      operationId: updateFinancement
      tags: [financement]
      summary: Mettre à jour un financement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFinancementBody"
      responses:
        "200":
          description: Financement mis à jour
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Financement"
    delete:
      operationId: deleteFinancement
      tags: [financement]
      summary: Supprimer un financement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Financement supprimé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /devis:
    get:
      operationId: listDevis
      tags: [devis]
      summary: Liste des devis de construction
      responses:
        "200":
          description: Devis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DevisConstruction"
    put:
      operationId: updateDevis
      tags: [devis]
      summary: Mettre à jour les devis de construction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDevisBody"
      responses:
        "200":
          description: Devis mis à jour
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DevisConstruction"
  /devis/puits-items:
    get:
      operationId: listPuitsItems
      tags: [devis]
      summary: Liste des items du puits
      responses:
        "200":
          description: Items du puits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PuitsItem"
    post:
      operationId: createPuitsItem
      tags: [devis]
      summary: Créer un item du puits
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePuitsItemBody"
      responses:
        "201":
          description: Item créé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PuitsItem"
  /devis/puits-items/{id}:
    put:
      operationId: updatePuitsItem
      tags: [devis]
      summary: Mettre à jour un item du puits
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePuitsItemBody"
      responses:
        "200":
          description: Item mis à jour
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PuitsItem"
    delete:
      operationId: deletePuitsItem
      tags: [devis]
      summary: Supprimer un item du puits
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Item supprimé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /financement/remboursements:
    get:
      operationId: getRemboursements
      tags: [financement]
      summary: Liste des remboursements
      responses:
        "200":
          description: Remboursements et soldes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RemboursementsResponse"
    post:
      operationId: createRemboursement
      tags: [financement]
      summary: Enregistrer un remboursement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRemboursementBody"
      responses:
        "201":
          description: Remboursement créé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Remboursement"
  /financement/remboursements/{id}:
    delete:
      operationId: deleteRemboursement
      tags: [financement]
      summary: Supprimer un remboursement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Supprimé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    Financement:
      type: object
      properties:
        id:
          type: integer
        nom:
          type: string
        montant:
          type: number
        date:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - nom
        - montant
        - date
        - createdAt
    CreateFinancementBody:
      type: object
      properties:
        nom:
          type: string
        montant:
          type: number
        date:
          type: string
          format: date
      required:
        - nom
        - montant
        - date
    DevisConstruction:
      type: object
      properties:
        id:
          type: integer
        batimentEstime:
          type: number
        batimentNotes:
          type: string
        carburantEstime:
          type: number
        puitsItems:
          type: array
          items:
            $ref: "#/components/schemas/PuitsItem"
        totalPuits:
          type: number
        totalGeneral:
          type: number
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - batimentEstime
        - carburantEstime
        - puitsItems
        - totalPuits
        - totalGeneral
        - updatedAt
    UpdateDevisBody:
      type: object
      properties:
        batimentEstime:
          type: number
        batimentNotes:
          type: string
        carburantEstime:
          type: number
    PuitsItem:
      type: object
      properties:
        id:
          type: integer
        designation:
          type: string
        quantite:
          type: number
        prixUnitaire:
          type: number
        prixTotal:
          type: number
      required:
        - id
        - designation
        - quantite
        - prixUnitaire
        - prixTotal
    CreatePuitsItemBody:
      type: object
      properties:
        designation:
          type: string
        quantite:
          type: number
        prixUnitaire:
          type: number
      required:
        - designation
        - quantite
        - prixUnitaire
...
    Remboursement:
      type: object
      properties:
        id:
          type: integer
        investisseurNom:
          type: string
        montant:
          type: number
        date:
          type: string
        commentaire:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, investisseurNom, montant, date]
    CreateRemboursementBody:
      type: object
      properties:
        investisseurNom:
          type: string
        montant:
          type: number
        date:
          type: string
        commentaire:
          type: string
      required: [investisseurNom, montant, date]
    SoldeInvestisseur:
      type: object
      properties:
        nom:
          type: string
        totalInvesti:
          type: number
        totalRembourse:
          type: number
        soldeRestant:
          type: number
      required: [nom, totalInvesti, totalRembourse, soldeRestant]
    RemboursementsResponse:
      type: object
      properties:
        remboursements:
          type: array
          items:
            $ref: "#/components/schemas/Remboursement"
        soldesInvestisseurs:
          type: array
          items:
            $ref: "#/components/schemas/SoldeInvestisseur"
      required: [remboursements, soldesInvestisseurs]
...